Move struct definitions to header files
